Separate checkpointing scenerio tests from preexisting checkpoint tests
